requirements.txt
# Core Runtime
python>=3.10,<3.13
pip>=23.0

# Async & Concurrency
asyncio==3.4.3
aiohttp>=3.9.0,<4.0
aiofiles>=23.2.1
uvloop>=0.19.0; sys_platform != "win32"

# Configuration & Dependency Management
pydantic>=2.5.0,<3.0
pyyaml>=6.0
dependency-injector>=4.41.0

# AI & Machine Learning
torch>=2.1.0,<2.3.0
transformers>=4.35.0,<5.0
sentence-transformers>=2.2.2
openai>=1.10.0
anthropic>=0.20.0
google-generativeai>=0.4.0
llama-cpp-python>=0.2.57  # For local Llama models
whisper>=1.1.10           # OpenAI Whisper for transcription
diffusers>=0.25.0         # For multimodal (if needed)
bitsandbytes>=0.41.0      # For quantized models

# NLP & Multilingual Support
spacy>=3.7.0
language-tool-python>=2.7.1
polyglot>=16.7.4
langdetect>=1.0.9

# Security & Cryptography
cryptography>=41.0.0
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
pyotp>=2.9.0
tink>=1.8.0               # Google's cryptographic library
secure-smtplib>=1.0.0     # Secure email

# Web & API Integration
requests>=2.31.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
selenium>=4.15.0          # Fallback for JS-heavy platforms
playwright>=1.40.0        # Modern browser automation

# Payment Processing
stripe>=5.0.0
paypalrestsdk>=1.13.1
yoomoney>=0.3.0
web3>=6.10.0              # For crypto payments (ETH, etc.)

# Data Storage & Caching
redis>=5.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
aiocache>=0.12.3

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
datadog>=0.45.0
grafana-api>=1.5.0

# Logging & Error Handling
structlog>=23.2.0
sentry-sdk>=1.40.0
loguru>=0.7.2

# File & Media Processing
moviepy>=1.0.3
pydub>=0.25.1
Pillow>=10.1.0

# Utilities
tenacity>=8.2.3           # Retry logic for self-healing
schedule>=1.2.0
python-dotenv>=1.0.0
deepdiff>=6.7.1           # For config/state comparison
jsonschema>=4.20.0        # Config validation

# Testing (for internal use, not required in prod)
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx-mock>=0.10.0

# Optional: For blockchain smart contracts
eth-brownie>=1.20.0       # Only if blockchain_service is active